unnecessary cast to float in SquadHealthHUD
(cherry picked from commit 498f0e4dd7491c0d1fe40362407f1f0685137090)

+++ src/main/java/fr/alexdoru/megawallsenhancementsmod/gui/huds/SquadHealthHUD.java
@@ -105,9 +105,9 @@ public class SquadHealthHUD implements IRenderer {
                 if (flag) {
                     final EntityPlayer entityplayer = this.mc.theWorld.getPlayerEntityByUUID(gameprofile.getId());
                     this.mc.getTextureManager().bindTexture(networkplayerinfo.getLocationSkin());
-                    Gui.drawScaledCustomSizeModalRect(xDrawingPos, yDrawingPos, 8.0F, (float) 8, 8, 8, 8, 8, 64.0F, 64.0F);
+                    Gui.drawScaledCustomSizeModalRect(xDrawingPos, yDrawingPos, 8, 8, 8, 8, 8, 8, 64.0F, 64.0F);
                     if (entityplayer == null || entityplayer.isWearing(EnumPlayerModelParts.HAT)) {
-                        Gui.drawScaledCustomSizeModalRect(xDrawingPos, yDrawingPos, 40.0F, 8, 8, 8, 8, 8, 64.0F, 64.0F);
+                        Gui.drawScaledCustomSizeModalRect(xDrawingPos, yDrawingPos, 40, 8, 8, 8, 8, 8, 64.0F, 64.0F);
                     }
                     xDrawingPos += 9;
                 }
